updated klaviyo environment a little
