+++ Sources/KlaviyoCore/Models/PushBackground.swift
@@ -1,0 +1,27 @@
+//
+//  PushBackground.swift
+//
+//
+//  Created by Ajay Subramanya on 8/8/24.
+//
+
+import UIKit
+
+public enum PushBackground: String, Codable {
+    case available = "AVAILABLE"
+    case restricted = "RESTRICTED"
+    case denied = "DENIED"
+
+    static func create(from status: UIBackgroundRefreshStatus) -> PushBackground {
+        switch status {
+        case .available:
+            return PushBackground.available
+        case .restricted:
+            return PushBackground.restricted
+        case .denied:
+            return PushBackground.denied
+        @unknown default:
+            return PushBackground.available
+        }
+    }
+}
